refactor: 예외 핸들러 한 곳으로 모음
